+++ java/TestNG_plugin/src/main/java/taurustestng/Utils.java
@@ -1,0 +1,13 @@
+package taurustestng;
+
+import java.io.PrintWriter;
+import java.io.StringWriter;
+
+public class Utils {
+    public static String getStackTrace(final Throwable throwable) {
+        final StringWriter sw = new StringWriter();
+        final PrintWriter pw = new PrintWriter(sw, true);
+        throwable.printStackTrace(pw);
+        return sw.getBuffer().toString();
+    }
+}