+++ site/dat/docs/Changelog.md
@@ -17,6 +17,7 @@
  - send `additional-classpath` jars into cloud when preparing a cloud test
  - fix proxy2jmx proxy settings for Chrome 
  - fix issue when user has too many tests and test is not found in BZA
+ - add TestNG-based runner for Selenium tests
 
 ## 1.7.2 <sup>13 oct 2016</sup>
  - fix keep-alive processing in Gatling